Workaround dependent type behavior
In order to keep old behavior for the dependent types the following
workaround is implemented. When matching two types expected and actual,
variables in actual type doesn't match the meta-types from expected atom.
In order to implement this all variables in actual type are
replaced by VariableMatch grounded atom which ignores meta-types passed.
For the usual type it returns the type binding of the original variable.