Optimize search for a case of the atom which is not in index storage
Before this change such case was handled as a custom matching. Which
means all atoms are extracted from index and matched with an atom from
query. Now we represent this case as en exact matching which is not in
the index. Thus only custom matchers for the specific branch of the
index are applied.